Remove all wrong code indentation
