+++ app/assets/stylesheets/application.css
@@ -13,37 +13,3 @@
  *= require_tree .
  *= require_self
  */
-.mynavbar {
-    float: right;
-}
-.main {
-    float: right;
-    font-size: 1.0rem;
-    margin-left: 5px;
-}
-.back {
-    float: right;
-    font-size: 1.0rem;
-}
-.mybtn {
-    float: right;
-}
-.comments {
-    border: solid;
-    width: auto;
-}
-.btnRegister {
-    float: left;
-    margin-top: 5px;
-    margin-left: 400px;
-}
-.user-section {
-    float: right;
-}
-.error-section {
-  width: 450px;
-  border: 2px solid red;
-  padding: 7px 7px 0;
-  margin-bottom: 20px;
-  background-color: #f0f0f0
-}
