+++ app/controllers/admin/base_controller.rb
@@ -1,0 +1,9 @@
+module Admin 
+  class BaseController < ApplicationController
+    before_action :validate_admin
+    
+    def validate_admin
+      # return redirect_to root_path unless current_user.admin?
+    end
+  end
+end
