+++ app/views/admin/users/index.html.erb
@@ -15,9 +15,9 @@
           <td><%= user.last_name %></td>
           <td><%= user.created_at%></td>
           <td><%= user.email %></td>
-          <td><%= link_to 'View Project', user , class:"btn btn-secondary" %></td>
-          <td> <%= link_to 'Edit', edit_user_path(user) %></td>
-          <td><%= link_to 'Delete', user, method: :delete, data: { confirm: 'Are you sure?' } %></td>
+          <td><%= link_to 'View Project', admin_user_path(user) , class:"btn btn-secondary" %></td>
+          <td> <%= link_to 'Edit', edit_admin_user_path(user) %></td>
+          <td><%= link_to 'Delete', admin_user_path(user), method: :delete, data: { confirm: 'Are you sure?' } %></td>
         </tr>
       <% end %>
       All users : <%= @user.count%>
