+++ app/views/admin/users/show.html.erb
@@ -24,5 +24,5 @@
 			</tr>
 		</tbody>
 	</table>
-  <%= link_to 'Back To Users', users_path, class: "btn btn-success"%>
+  <%= link_to 'Back To Users', admin_users_path, class: "btn btn-success"%>
 </div>
