+++ app/views/layouts/_navbar.html.erb
@@ -13,7 +13,7 @@
           <a class="nav-link active"><%= link_to 'Edit Profile', edit_user_registration_path, class: "btn btn-outline-dark main" %></a>
         </li>
         <li class="nav-item">
-          <a class="nav-link active"><%= link_to 'User List', users_path, class: "btn btn-outline-dark main" %></a>
+          <a class="nav-link active"><%= link_to 'User List', admin_users_path, class: "btn btn-outline-dark main" %></a>
         </li>
       </ul>
       <% else %>
