+++ app/views/projects/show.html.erb
@@ -1,7 +1,7 @@
 	<h1 class="text-center">Project <%= @project.name %></h1>
 	<p id="notice"><%= notice %></p>
 	<%= link_to 'Go Back', projects_path, class: "btn btn-secondary mybtn" %>
-	<%= link_to "Create User", new_user_path, class: "btn btn-secondary mybtn" %>
+	<%= link_to "Create User", new_admin_user_path, class: "btn btn-secondary mybtn" %>
 <div class="container">	
 	<table class="">
 		<tbody>
