+++ config/routes.rb
@@ -1,10 +1,15 @@
 Rails.application.routes.draw do
-  resources :clients
+
   devise_for :users
-  resources :users
-  root to: 'projects#index'
+
+  namespace :admin  do 
+    resources :users
+end 
+
   resources :projects do
     resources :comments
-  end
-  
+end
+  resources :clients
+
+  root to: 'projects#index' 
 end
